requirements.txt
asgiref==3.8.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
django-localflavor==4.0
django-parler==2.3
django-rosetta==0.10.0
django-taggit==6.0.0
Django==4.2.11
flower==2.0.1
fonttools==4.51.0
humanize==4.9.0
pillow==10.2.0
polib==1.2.0
prometheus_client==0.20.0
python-dateutil==2.9.0.post0
python-stdnum==1.20
redis==5.0.4
requests==2.31.0
stripe==9.3.0
weasyprint==61.2
zopfli==0.2.3
